LPS-107862 Only copy over script files if they exist

+++ modules/sdk/gradle-plugins-workspace/src/main/resources/com/liferay/gradle/plugins/workspace/configurators/100_liferay_image_setup.sh.tpl
@@ -18,16 +18,28 @@ function copy_and_remove_scripts {
 	echo "[LIFERAY] Copying /home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts files:"
 	echo ""
 
+	if [ -d "/home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts)" ] && [ "$(ls -A /home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts)" ]; then
+
 	tree --noreport /home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts
+
+	fi
 
 	echo ""
 	echo "[LIFERAY] ... into ${LIFERAY_MOUNT_DIR}/scripts"
 
 	mkdir -p ${LIFERAY_MOUNT_DIR}/scripts
 
+	if [ -d "/home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts)" ] && [ "$(ls -A /home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts)" ]; then
+
 	cp -r /home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts/* ${LIFERAY_MOUNT_DIR}/scripts
 
+	fi
+
+	if [ -d "/home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts)" ]; then
+
 	rm -rf /home/liferay/configs/${LIFERAY_WORKSPACE_ENVIRONMENT}/scripts
+
+	fi
 
 	echo ""
 }
